+++ Client/package.json
@@ -18,6 +18,7 @@
     "cors": "^2.8.5",
     "cross-fetch": "^3.0.6",
     "d3": "^5.5.0",
+    "email-validator": "^2.0.4",
     "enzyme-matchers": "^7.1.2",
     "formik": "^2.2.6",
     "immutable": "^4.0.0-rc.12",
