+++ Client/src/Authentication/SignUp.tsx
@@ -1,0 +1,205 @@
+import React, { Fragment, useState } from 'react';
+import Avatar from '@material-ui/core/Avatar';
+import Button from '@material-ui/core/Button';
+import CssBaseline from '@material-ui/core/CssBaseline';
+import FormControlLabel from '@material-ui/core/FormControlLabel';
+import Checkbox from '@material-ui/core/Checkbox';
+import Link from '@material-ui/core/Link';
+import Grid from '@material-ui/core/Grid';
+import Box from '@material-ui/core/Box';
+import LockOutlinedIcon from '@material-ui/icons/LockOutlined';
+import Typography from '@material-ui/core/Typography';
+import { makeStyles } from '@material-ui/core/styles';
+import Container from '@material-ui/core/Container';
+import { FastField, Form, Formik } from 'formik';
+import { signupValidationSchema } from './AuthenticationValidationSchema';
+import { IPasswordSettings, ISignUpUser } from '../Models/Profile/IProfileModel';
+import { MuiTextFieldFormik } from '../Components/Forms/FormikFields';
+
+function Copyright() {
+  return (
+    <Typography variant="body2" color="textSecondary" align="center">
+      {'Copyright © '}
+      <Link color="inherit" href="https://material-ui.com/">
+        Your Website
+      </Link>{' '}
+      {new Date().getFullYear()}
+      {'.'}
+    </Typography>
+  );
+}
+
+const useStyles = makeStyles((theme) => ({
+  paper: {
+    marginTop: theme.spacing(8),
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center',
+  },
+  avatar: {
+    margin: theme.spacing(1),
+    backgroundColor: theme.palette.secondary.main,
+  },
+  form: {
+    width: '100%', // Fix IE 11 issue.
+    marginTop: theme.spacing(3),
+  },
+  submit: {
+    margin: theme.spacing(3, 0, 2),
+  },
+}));
+
+export default function SignUp() {
+  const classes = useStyles();
+
+  const [email, setEmail] = useState("");
+
+  const user: ISignUpUser = {
+    firstName: "",
+    lastName: "",
+    organization: "",
+    email: "",
+    password: "",
+    confirmPassword: ""
+  }
+
+  const handleSignUpSubmit = (user: ISignUpUser): void => {
+    console.log(JSON.stringify(user))
+    // TODO: hook up to the backend
+  }
+
+  return (
+    <Fragment>
+      <Container component="main" maxWidth="xs">
+        <CssBaseline />
+        <div className={classes.paper}>
+          <Avatar className={classes.avatar}>
+            <LockOutlinedIcon />
+          </Avatar>
+          <Typography component="h1" variant="h5">
+            Sign up
+        </Typography>
+          <Formik
+            initialValues={user}
+            validationSchema={signupValidationSchema}
+            onSubmit={handleSignUpSubmit}
+          >
+            {props => {
+              const {
+                values,
+                touched,
+                errors,
+                isSubmitting,
+                handleChange,
+                handleBlur,
+                handleSubmit
+              } = props;
+              return (
+                <>
+                  <Form className={classes.form} noValidate>
+                    <Grid container spacing={2}>
+                      <Grid item xs={12} sm={6}>
+                        <FastField
+                          autoComplete="fname"
+                          name="firstName"
+                          label="First Name"
+                          required
+                          component={MuiTextFieldFormik}
+                          id="firstName"
+                          autoFocus
+                          variant="outlined"
+                          fullWidth />
+                      </Grid>
+                      <Grid item xs={12} sm={6}>
+                        <FastField
+                          variant="outlined"
+                          required
+                          fullWidth
+                          component={MuiTextFieldFormik}
+                          id="lastName"
+                          label="Last Name"
+                          name="lastName"
+                          autoComplete="lname" />
+                      </Grid>
+                      <Grid item xs={12}>
+                        <FastField
+                          variant="outlined"
+                          required
+                          fullWidth
+                          component={MuiTextFieldFormik}
+                          id="email"
+                          label="Email Address"
+                          name="email"
+                          autoComplete="email" />
+                      </Grid>
+                      <Grid item xs={12}>
+                        <FastField
+                          variant="outlined"
+                          required
+                          fullWidth
+                          component={MuiTextFieldFormik}
+                          id="organization"
+                          label="Organization"
+                          name="organization"
+                          autoComplete="organization" />
+                      </Grid>
+                      <Grid item xs={12}>
+                        <FastField
+                          variant="outlined"
+                          required
+                          fullWidth
+                          component={MuiTextFieldFormik}
+                          name="password"
+                          label="Password"
+                          type="password"
+                          id="password"
+                          autoComplete="current-password" />
+                      </Grid>
+                      <Grid item xs={12}>
+                        <FastField
+                          variant="outlined"
+                          required
+                          fullWidth
+                          component={MuiTextFieldFormik}
+                          name="confirmPassword"
+                          label="Confirm Password"
+                          type="password"
+                          id="confirmPassword"
+                          autoComplete="confirm-password" />
+                      </Grid>
+                      <Grid item xs={12}>
+                        <FormControlLabel
+                          control={<Checkbox value="allowExtraEmails" color="primary" />}
+                          label="I want to receive inspiration, marketing promotions and updates via email." />
+                      </Grid>
+                    </Grid>
+                    <Button
+                      type="submit"
+                      fullWidth
+                      variant="contained"
+                      color="primary"
+                      className={classes.submit}
+                    >
+                      Sign Up
+                    </Button>
+                    <Grid container justify="flex-end">
+                      <Grid item>
+                        <Link href="#" variant="body2">
+                          Already have an account?Sign in
+                        </Link>
+                      </Grid>
+                    </Grid>
+                  </Form>
+                </>
+              );
+            }}
+          </Formik>
+        </div>
+        <Box mt={5}>
+          <Copyright />
+        </Box>
+      </Container>
+    </Fragment >
+  );
+}
+
